+++ src/nsp/azext_nsp/aaz/latest/network/perimeter/_list.py
@@ -181,7 +181,9 @@ class List(AAZCommand):
             )
             _element.location = AAZStrType()
             _element.name = AAZStrType()
-            _element.properties = AAZObjectType()
+            _element.properties = AAZObjectType(
+                flags={"client_flatten": True},
+            )
             _element.tags = AAZDictType()
             _element.type = AAZStrType(
                 flags={"read_only": True},
@@ -295,7 +297,9 @@ class List(AAZCommand):
             )
             _element.location = AAZStrType()
             _element.name = AAZStrType()
-            _element.properties = AAZObjectType()
+            _element.properties = AAZObjectType(
+                flags={"client_flatten": True},
+            )
             _element.tags = AAZDictType()
             _element.type = AAZStrType(
                 flags={"read_only": True},