+++ src/nsp/azext_nsp/aaz/latest/network/perimeter/onboarded_resources/_list.py
@@ -159,7 +159,9 @@ class List(AAZCommand):
             )
             _element.location = AAZStrType()
             _element.name = AAZStrType()
-            _element.properties = AAZObjectType()
+            _element.properties = AAZObjectType(
+                flags={"client_flatten": True},
+            )
             _element.tags = AAZDictType()
             _element.type = AAZStrType(
                 flags={"read_only": True},